requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1
pandas==2.2.2
numpy==1.26.4
pvlib==0.10.4
joblib==1.5.1
matplotlib==3.10.3
scikit-learn==1.6.1
scipy==1.15.3

# Utils
django-debug-toolbar==4.2.0
Pillow==10.0.1

# Services
celery==5.3.4
redis==5.0.1
hiredis==2.2.3

# Tooling
django-celery-results==2.5.1
djangorestframework==3.14.0
django-api-generator==1.0.17
drf-spectacular==0.26.5

# Deployment
whitenoise==6.5.0
gunicorn==21.2.0

# DB Layer
# psycopg2-binary==3.1.12
# mysqlclient==2.2.0
